ref(ts): Convert sessionHistory to typescript
