+++ src/sentry/static/sentry/app/views/settings/account/accountSecurity/sessionHistory/utils.tsx
@@ -1,0 +1,8 @@
+import space from 'app/styles/space';
+
+export const tableLayout = `
+  display: grid;
+  grid-template-columns: auto 140px 140px;
+  grid-gap ${space(1)};
+  align-items: center;
+`;
